requirements.txt
web3[tester]
py-solc-x